requirements.txt
Django==5.2.7
djangorestframework==3.16.1
mysqlclient==2.2.7
python-dotenv==1.1.1
requests==2.32.5
django-filter==25.2
